+++ src/main/java/de/foursoft/discordbot/BotInitializer.java
@@ -1,12 +1,8 @@
 package de.foursoft.discordbot;
 
 import com.jagrosh.jdautilities.commons.waiter.EventWaiter;
-import de.foursoft.discordbot.commands.PingCommand;
-import de.foursoft.discordbot.commands.ReactCommand;
-import de.foursoft.discordbot.commands.ResetCommand;
-import de.foursoft.discordbot.commands.SecretCommand;
-import net.dv8tion.jda.api.JDA;
 import net.dv8tion.jda.api.JDABuilder;
+import net.dv8tion.jda.api.hooks.EventListener;
 import net.dv8tion.jda.api.requests.GatewayIntent;
 import net.dv8tion.jda.api.utils.ChunkingFilter;
 import net.dv8tion.jda.api.utils.MemberCachePolicy;
@@ -15,32 +11,31 @@ import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
+import javax.annotation.PostConstruct;
 import javax.security.auth.login.LoginException;
 
 @Component
 public class BotInitializer {
     private static final Logger LOGGER = LoggerFactory.getLogger(FourSoftDiscordBot.class);
 
-    private final JDA discordApi;
-
     @Autowired
     private EventWaiter eventWaiter;
     @Autowired
-    private ListenerAdapterDispatcher dispatcher;
+    private EventListener eventListener;
 
-    public BotInitializer() {
+    @PostConstruct
+    private void init(){
         Configuration config = new Configuration("config.properties");
 
-        JDA tmpJda = null;
         try {
-            tmpJda = JDABuilder.createDefault(config.getToken())  // init default settings
+            JDABuilder.createDefault(config.getToken())  // init default settings
                     .setChunkingFilter(ChunkingFilter.ALL)  // load all members
                     .enableIntents(
                             GatewayIntent.GUILD_MEMBERS, // allow access to guild members
                             GatewayIntent.GUILD_PRESENCES  // access to online status and activities
                     )
                     .setMemberCachePolicy(MemberCachePolicy.ALL)  // always cache members
-                    .addEventListeners(dispatcher)
+                    .addEventListeners(eventListener)
                     .addEventListeners(eventWaiter)  // waits for events
                     .build().awaitReady();  // wait until the API is ready
         } catch (LoginException e) {
@@ -50,11 +45,5 @@ public class BotInitializer {
             LOGGER.error("Could not start the Bot, interrupted while waiting!");
             System.exit(-1);
         }
-
-        discordApi = tmpJda;
-    }
-
-    public JDA getDiscordApi() {
-        return discordApi;
     }
 }
