+++ src/main/java/de/foursoft/discordbot/ListenerAdapterDispatcher.java
@@ -1,19 +1,32 @@
 package de.foursoft.discordbot;
 
-import de.foursoft.discordbot.commands.Command;
+import de.foursoft.discordbot.eventconsumer.EventConsumer;
 import net.dv8tion.jda.api.events.Event;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
+import javax.annotation.PostConstruct;
 import java.lang.reflect.InvocationHandler;
 import java.lang.reflect.Method;
+import java.util.Collections;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
+import java.util.stream.Collectors;
 
 @Component
 public class ListenerAdapterDispatcher implements InvocationHandler {
 
     @Autowired
-    private List<Command> allCommands;
+    private List<EventConsumer> allEventConsumers;
+
+    private Map<Class, List<EventConsumer>> consumerByEvent;
+
+    @PostConstruct
+    private void init(){
+        consumerByEvent = allEventConsumers.stream().collect(Collectors.groupingBy(EventConsumer::getClassOfT));
+    }
+
 
     @Override
     public Object invoke(Object proxy, Method method, Object[] args) throws Throwable {
@@ -24,8 +37,7 @@ public class ListenerAdapterDispatcher implements InvocationHandler {
 
         Object event = args[0];
 
-        allCommands.stream()
-                .filter(c -> c.getClassOfT().equals(event.getClass()))
+        consumerByEvent.getOrDefault(event.getClass(), Collections.emptyList())
                 .forEach(c -> c.execute((Event) event));
 
         return null;
