+++ src/main/java/de/foursoft/discordbot/eventconsumer/ReactEventConsumer.java
@@ -1,4 +1,4 @@
-package de.foursoft.discordbot.commands;
+package de.foursoft.discordbot.eventconsumer;
 
 import net.dv8tion.jda.api.entities.Message;
 import net.dv8tion.jda.api.events.message.guild.GuildMessageReceivedEvent;
@@ -7,8 +7,8 @@ import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Component;
 
 @Component
-public class ReactCommand extends GuildMessageReceivedCommand {
-    private static final Logger LOGGER = LoggerFactory.getLogger(ReactCommand.class);
+public class ReactEventConsumer extends GuildMessageReceivedEventConsumer {
+    private static final Logger LOGGER = LoggerFactory.getLogger(ReactEventConsumer.class);
 
     private static final String THUMBS_UP_UNICODE = "\uD83D\uDC4D";
